requirements.txt
adrf==0.1.6
annotated-types==0.6.0
asgiref==3.8.1
async-property==0.2.2
click==8.1.7
Django==5.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-ninja==1.1.0
django-phonenumber-field==7.3.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==24.11.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
phonenumberslite==8.13.35
pluggy==1.5.0
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
